Subtitutions
Does not compile, but want to have the version in the history.

+++ src/camera.cpp
@@ -8,31 +8,31 @@ using namespace glm;
 
 // By default, the camera is at (0,0,0,1) and looking in the
 // negative z-direction (0,0,-1,0).
-Camera::Camera(float fov, float ratio, float near, float far)
+Camera::Camera(double fov, double ratio, double near, double far)
 {
     _fov = fov;
     _ratio = ratio;
     _near = near;
     _far = far;
     _proj = glm::perspective(fov, ratio, near, far);
-    _view = glm::mat4(); //defaults to identity matrix
+    _view = glm::dmat4(); //defaults to identity matrix
 }
 
 // Return the camera's current projection matrix.
-glm::mat4 Camera::get_projection()
+glm::dmat4 Camera::get_projection()
 {
     return _proj;
 }
 
 // Return the camera's current view matrix.
-glm::mat4 Camera::get_view()
+glm::dmat4 Camera::get_view()
 {
     return _view;
 }
 
 // Change the camera's aspect ratio. Useful when resizing
 // the window.
-void Camera::set_ratio(float ratio)
+void Camera::set_ratio(double ratio)
 {
     if(_ratio == ratio)
         return;
@@ -47,7 +47,7 @@ void Camera::set_ratio(float ratio)
 // formation you want to apply. This avoids explicitly inverting
 // the matrix, which is inferior to generating the inverse from the
 // start.
-void Camera::transform(glm::mat4 transformationinv)
+void Camera::transform(glm::dmat4 transformationinv)
 {
     _view =  transformationinv * _view;
 }
@@ -55,7 +55,7 @@ void Camera::transform(glm::mat4 transformationinv)
 // Sets the transformation of the camera.
 // The matrix supplied must be the *inverse* of the trans-
 // formation you want.
-void Camera::set_transformation(glm::mat4 transformationinv)
+void Camera::set_transformation(glm::dmat4 transformationinv)
 {
     _view = transformationinv;
 }
