+++ src/cameracontrols.cpp
@@ -21,7 +21,7 @@ CameraControls::CameraControls(GLFWwindow* window, Camera* camera, ovrHmd* hmd)
     reset_to_origin();
 }
 
-void CameraControls::handle(float delta_time, int width, int height)
+void CameraControls::handle(double delta_time, int width, int height)
 {
     handle_keyboard(delta_time);
     if(bind_mouse)
@@ -34,7 +34,7 @@ void CameraControls::handle(float delta_time, int width, int height)
     update_camera_transformation();
 }
 
-void CameraControls::handle_mouse(float delta_time, int width, int height)
+void CameraControls::handle_mouse(double delta_time, int width, int height)
 {
     double mouse_x, mouse_y;
     double center_x = floor(width/2);
@@ -45,11 +45,11 @@ void CameraControls::handle_mouse(float delta_time, int width, int height)
     glfwSetCursorPos(_window, center_x, center_y);
 
     double _angle_ver = _mouse_speed * delta_time * (center_y - mouse_y);
-    _head.up = ((float)cos(_angle_ver))*_head.up - ((float)sin(_angle_ver))*_head.forward;
+    _head.up = ((double)cos(_angle_ver))*_head.up - ((double)sin(_angle_ver))*_head.forward;
 
     double angle_hor = _mouse_speed * delta_time * (center_x - mouse_x);
-    _shoulders.forward = ((float)cos(angle_hor))*_shoulders.forward - ((float)sin(angle_hor))*_shoulders.right;
-    _shoulders.right = ((float)sin(angle_hor))*_shoulders.forward + ((float)cos(angle_hor))*_shoulders.right;
+    _shoulders.forward = ((double)cos(angle_hor))*_shoulders.forward - ((double)sin(angle_hor))*_shoulders.right;
+    _shoulders.right = ((double)sin(angle_hor))*_shoulders.forward + ((double)cos(angle_hor))*_shoulders.right;
 
     _head.forward = _shoulders.forward;
     _head.right = _shoulders.right;
@@ -69,7 +69,7 @@ void CameraControls::handle_hmd()
 
 void CameraControls::update_camera_transformation()
 {
-    glm::mat4 transf = glm::mat4();
+    glm::dmat4 transf = glm::dmat4();
     transf[0][0] = _head.right.x;
     transf[0][1] = _head.right.y;
     transf[0][2] = _head.right.z;
@@ -94,10 +94,10 @@ void CameraControls::update_camera_transformation()
     _camera->set_transformation(transf);
 }
 
-void CameraControls::move_right(float distance)
+void CameraControls::move_right(double distance)
 {
-    glm::vec4 newpos = hypermath::exp(_shoulders.pos, distance * _head.right);
-    glm::mat4 transf = hypermath::translation(_shoulders.pos,newpos);
+    glm::dvec4 newpos = hypermath::exp(_shoulders.pos, distance * _head.right);
+    glm::dmat4 transf = hypermath::translation(_shoulders.pos,newpos);
     _shoulders.pos = newpos;
     _shoulders.right = transf * _shoulders.right;
     _shoulders.forward = transf * _shoulders.forward;
@@ -107,33 +107,33 @@ void CameraControls::move_right(float distance)
     update_camera_transformation();
 }
 
-void CameraControls::handle_keyboard(float delta_time)
+void CameraControls::handle_keyboard(double delta_time)
 {
     //direction of walking in coordinates (right, up, forward)
-    glm::vec3 walking_direction = glm::vec3();
+    glm::dvec3 walking_direction = glm::dvec3();
     if( glfwGetKey(_window, GLFW_KEY_UP) || glfwGetKey(_window, GLFW_KEY_W ))
     {
-        walking_direction += glm::vec3(0,0,1);
+        walking_direction += glm::dvec3(0,0,1);
     }
     if( glfwGetKey(_window, GLFW_KEY_DOWN) || glfwGetKey(_window, GLFW_KEY_S ))
     {
-        walking_direction += glm::vec3(0,0,-1);
+        walking_direction += glm::dvec3(0,0,-1);
     }    
     if( glfwGetKey(_window, GLFW_KEY_LEFT) || glfwGetKey(_window, GLFW_KEY_A ))
     {
-        walking_direction += glm::vec3(-1,0,0);
+        walking_direction += glm::dvec3(-1,0,0);
     }
     if( glfwGetKey(_window, GLFW_KEY_RIGHT) || glfwGetKey(_window, GLFW_KEY_D ))
     {
-        walking_direction += glm::vec3(1,0,0);
+        walking_direction += glm::dvec3(1,0,0);
     }
     if( glfwGetKey(_window, GLFW_KEY_PAGE_UP) || glfwGetKey(_window, GLFW_KEY_R))
     {
-        walking_direction += glm::vec3(0,1,0);
+        walking_direction += glm::dvec3(0,1,0);
     }
     if( glfwGetKey(_window, GLFW_KEY_PAGE_DOWN) || glfwGetKey(_window, GLFW_KEY_F))
     {
-        walking_direction += glm::vec3(0,-1,0);
+        walking_direction += glm::dvec3(0,-1,0);
     }
     if ( glfwGetKey(_window, GLFW_KEY_GRAVE_ACCENT))
     {
@@ -141,13 +141,13 @@ void CameraControls::handle_keyboard(float delta_time)
     }
 
     //length is either zero or at least one, mathematically
-    //but floating point
+    //but doubleing point
     if(glm::length(walking_direction) >= 0.5)
     {
         walking_direction = glm::normalize(walking_direction);
         walking_direction *= _move_speed * delta_time * _meter;
-        glm::vec4 newpos = hypermath::exp(_shoulders.pos, walking_direction.x*_shoulders.right + walking_direction.y*_shoulders.up + walking_direction.z*_shoulders.forward);
-        glm::mat4 transf = hypermath::translation(_shoulders.pos,newpos);
+        glm::dvec4 newpos = hypermath::exp(_shoulders.pos, walking_direction.x*_shoulders.right + walking_direction.y*_shoulders.up + walking_direction.z*_shoulders.forward);
+        glm::dmat4 transf = hypermath::translation(_shoulders.pos,newpos);
         _shoulders.pos = newpos;
 
         // This makes movement weird 
@@ -161,56 +161,56 @@ void CameraControls::handle_keyboard(float delta_time)
 void CameraControls::reset_to_origin()
 {
 
-    _shoulders.pos = glm::vec4(0,0,0,1);
-    _shoulders.forward = glm::vec4(0,0,-1,0);
-    _shoulders.right = glm::vec4(1,0,0,0);
-    _shoulders.up = glm::vec4(0,1,0,0);
+    _shoulders.pos = glm::dvec4(0,0,0,1);
+    _shoulders.forward = glm::dvec4(0,0,-1,0);
+    _shoulders.right = glm::dvec4(1,0,0,0);
+    _shoulders.up = glm::dvec4(0,1,0,0);
 
-    _head.pos = glm::vec4(0,0,0,1);
-    _head.forward = glm::vec4(0,0,-1,0);
-    _head.right = glm::vec4(1,0,0,0);
-    _head.up = glm::vec4(0,1,0,0);
+    _head.pos = glm::dvec4(0,0,0,1);
+    _head.forward = glm::dvec4(0,0,-1,0);
+    _head.right = glm::dvec4(1,0,0,0);
+    _head.up = glm::dvec4(0,1,0,0);
 }
 
 void CameraControls::set_orientation(glm::quat rotation)
 {
-    glm::mat4 rotate = hypermath::rotation0(rotation);
-    glm::mat4 rotatez = hypermath::rotationz(rotation);
-    glm::mat4 transf = hypermath::translation(glm::vec4(0,0,0,1),_shoulders.pos);
-    _head.up = transf*rotate*glm::vec4(0,1,0,0);
-    _head.forward = transf*rotate*glm::vec4(0,0,-1,0);
-    _head.right = transf*rotate*glm::vec4(1,0,0,0);
-    _shoulders.forward = transf*rotatez*glm::vec4(0,0,-1,0);
-    _shoulders.right = transf*rotatez*glm::vec4(1,0,0,0);
+    glm::dmat4 rotate = hypermath::rotation0(rotation);
+    glm::dmat4 rotatez = hypermath::rotationz(rotation);
+    glm::dmat4 transf = hypermath::translation(glm::dvec4(0,0,0,1),_shoulders.pos);
+    _head.up = transf*rotate*glm::dvec4(0,1,0,0);
+    _head.forward = transf*rotate*glm::dvec4(0,0,-1,0);
+    _head.right = transf*rotate*glm::dvec4(1,0,0,0);
+    _shoulders.forward = transf*rotatez*glm::dvec4(0,0,-1,0);
+    _shoulders.right = transf*rotatez*glm::dvec4(1,0,0,0);
     update_camera_transformation();
 }
 
-void CameraControls::set_mouse_speed(float speed)
+void CameraControls::set_mouse_speed(double speed)
 {
     _mouse_speed = speed;
 }
 
-void CameraControls::set_step_size(float size)
+void CameraControls::set_step_size(double size)
 {
     _move_speed = size;
 }
 
-float CameraControls::get_ipd()
+double CameraControls::get_ipd()
 {
     return _ipd;
 }
 
-float CameraControls::get_height()
+double CameraControls::get_height()
 {
     return _height;
 }
 
-float CameraControls::get_meter()
+double CameraControls::get_meter()
 {
     return _meter;
 }
 
-glm::vec4 CameraControls::get_pos()
+glm::dvec4 CameraControls::get_pos()
 {
     return _shoulders.pos;
 }
@@ -218,7 +218,7 @@ glm::vec4 CameraControls::get_pos()
 //increase step size. max is 1.0f
 void CameraControls::increase_speed()
 {
-    float size = _move_speed + .05f;
+    double size = _move_speed + .05f;
     if(size > 1.0f)
     {
         size = 1.0f;
@@ -229,7 +229,7 @@ void CameraControls::increase_speed()
 //decrease step size. minimum is 0.01f
 void CameraControls::decrease_speed()
 {
-    float size = _move_speed - .05f;
+    double size = _move_speed - .05f;
     if(size <= 0)
     {
         size = 0.01f;
@@ -238,29 +238,29 @@ void CameraControls::decrease_speed()
     set_step_size(size);
 }
 
-void CameraControls::grow(float factor)
+void CameraControls::grow(double factor)
 {
     _meter *= factor;
 }
 
-void CameraControls::shrink(float factor)
+void CameraControls::shrink(double factor)
 {
     _meter /= factor;
 }
 
-glm::mat4 CameraControls::get_cam_view()
+glm::dmat4 CameraControls::get_cam_view()
 {
-    glm::mat4 view = _camera->get_view();
+    glm::dmat4 view = _camera->get_view();
     return view;
 }
 
-glm::vec4 CameraControls::get_forward()
+glm::dvec4 CameraControls::get_forward()
 {
-    glm::vec4 forward = _shoulders.forward;
+    glm::dvec4 forward = _shoulders.forward;
     return forward;
 }
 
-glm::vec4 CameraControls::get_flag_pos()
+glm::dvec4 CameraControls::get_flag_pos()
 {
     return hypermath::midpoint(_shoulders.forward, -_shoulders.up, 0.5f);
 }
