+++ src/primitives.h
@@ -7,20 +7,20 @@
 
 namespace primitives
 {
-    mesh line(glm::vec4 a, glm::vec4 b, glm::vec4 col);
-    mesh axes(float radius);
+    mesh line(glm::dvec4 a, glm::dvec4 b, glm::dvec4 col);
+    mesh axes(double radius);
     mesh grid(double grid_space);
     mesh gravity_field(double grid_space);
-    mesh triangle(glm::vec4 a, glm::vec4 b, glm::vec4 c);
-    mesh rectangle(float width, float depth, glm::vec4 color);
-    mesh circumscribed_ngon(int n, float radius, glm::vec4 color);
+    mesh triangle(glm::dvec4 a, glm::dvec4 b, glm::dvec4 c);
+    mesh rectangle(double width, double depth, glm::dvec4 color);
+    mesh circumscribed_ngon(int n, double radius, glm::dvec4 color);
 
-    mesh tetrahedron(float radius);
-    mesh octahedron(float radius);
-    mesh object(std::vector<glm::vec4> v);
-    mesh subdivided_triangle(glm::vec4 a, glm::vec4 b, glm::vec4 c, int divisions, bool sierpinski = false);
-    mesh subdivided_octahedron(float radius, int divisions, bool sierpinski = false);
-    mesh sphere(float radius, int divisions, glm::vec4 col, bool sierpinski = false);
+    mesh tetrahedron(double radius);
+    mesh octahedron(double radius);
+    mesh object(std::vector<glm::dvec4> v);
+    mesh subdivided_triangle(glm::dvec4 a, glm::dvec4 b, glm::dvec4 c, int divisions, bool sierpinski = false);
+    mesh subdivided_octahedron(double radius, int divisions, bool sierpinski = false);
+    mesh sphere(double radius, int divisions, glm::dvec4 col, bool sierpinski = false);
 }
 
 #endif
